+++ c/rpg/src/present.h
@@ -3,7 +3,8 @@
 #ifndef __PRESENT_H__
 #define __PRESENT_H__
 
-void show_presentation();
+void show_presentation(char *version);
 void show_header();
+void monster_attack_effect();
 
 #endif